+++ src/profile/games/mo6tw/musicmenu.cpp
@@ -65,7 +65,7 @@ void Configure() {
   FadeOutDuration = EnsureGetMember<float>("FadeOutDuration");
 
   auto drawType = Game::DrawComponentType::_from_integral_unchecked(
-      EnsureGetMember<int>("DrawType"));
+      EnsureGetMember<uint8_t>("DrawType"));
 
   UI::Menus[drawType].push_back(new UI::MO6TW::MusicMenu());
 }
